requirements.txt
pyusb
numpy
matplotlib
git+https://github.com/nmigen/nmigen.git
git+https://github.com/hansfbaier/usb2-highspeed-core.git
git+https://github.com/hansfbaier/nmigen-library.git
